requirements.txt
altgraph==0.17.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
duckdb==1.1.2
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.5.0
MouseInfo==0.1.3
mysql-connector-python==9.0.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMsgBox==1.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
SecretStorage==3.3.3
setuptools==75.3.0
shiboken6==6.8.1
six==1.16.0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.3
watchdog==6.0.0